conf

+++ Laboratorio 1/Ejercicio 1.py
@@ -1,7 +1,7 @@
 c=input("ingrese el primer número: ")
-c=int(c)
+c=float(c)
 b=input("ingrese el segundo número: ")
-b=int(b)
+b=float(b)
 def max_num(c,b):
     if c>b:
         return c
